Collision Complete

+++ StudioProject02_C++/Source/Player.cpp
@@ -1,7 +1,7 @@
 #include "Player.h"
 #include "Application.h"
 
-Player::Player(const Vector3& pos, const Vector3& dir, int rangeX, int rangeY, int rangeZ, bool isCollisionON, float mass, std::string name)
+Player::Player(const Vector3& pos, const Vector3& dir, float rangeX, float rangeY, float rangeZ, bool isCollisionON, float mass, std::string name)
 {
 	this->position = pos;
 	this->movingObj = false;
